requirements.txt
uagents>=0.14.0
anthropic>=0.39.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.10.0

# MCP (Model Context Protocol) Support
mcp>=0.9.0
